+++ tests/test_config.py
@@ -28,6 +28,7 @@ import copy
 import json_checker
 from transitions import MachineError
 
+import numpy as np
 import tests.common as common
 import pandora.check_json as JSON_checker
 from pandora.state_machine import PandoraMachine
@@ -395,6 +396,8 @@ class TestConfig(unittest.TestCase):
             "input": {
                 "nodata_left": -9999,
                 "nodata_right": -9999,
+                "img_left": "tests/pandora/left.png",
+                "img_right": "tests/pandora/right.png",
                 "left_mask": None,
                 "right_mask": None,
                 "left_classif": None,
@@ -403,8 +406,6 @@ class TestConfig(unittest.TestCase):
                 "right_segm": None,
                 "disp_min_right": None,
                 "disp_max_right": None,
-                "img_left": "tests/pandora/left.png",
-                "img_right": "tests/pandora/right.png",
                 "disp_min": -60,
                 "disp_max": 0,
             },
@@ -431,6 +432,42 @@ class TestConfig(unittest.TestCase):
         pandora_machine = PandoraMachine()
 
         self.assertRaises(MachineError, JSON_checker.check_pipeline_section, cfg_pipeline, pandora_machine)
+
+    @staticmethod
+    def test_memory_consumption_estimation():
+        """
+        Test the method test_memory_consumption_estimation
+        """
+
+        # Most consuming function is to_disp
+        cv_size = 450 * 375 * 61
+        m_line = 8.68e-06
+        n_line = 243
+        # Compute memory consumption in GiB with a marge of +-10%
+        consumption_vt = (
+            ((cv_size * m_line + n_line) * (1 - 0.1)) / 1024,
+            ((cv_size * m_line + n_line) * (1 + 0.1)) / 1024,
+        )
+
+        # Run memory_consumption_estimation giving the input parameters directly
+        img_left_path = "tests/pandora/left.png"
+        disp_min = -60
+        disp_max = 0
+        pandora_machine = PandoraMachine()
+        pipeline_cfg = {"pipeline": copy.deepcopy(common.basic_pipeline_cfg)}
+
+        min_mem_consump, max_mem_consump = JSON_checker.memory_consumption_estimation(
+            pipeline_cfg, (img_left_path, disp_min, disp_max), pandora_machine
+        )
+        np.testing.assert_allclose((min_mem_consump, max_mem_consump), consumption_vt, rtol=1e-02)
+
+        # Run memory_consumption_estimation giving the input parameters in a dict
+        pandora_machine = PandoraMachine()
+        input_cfg = {"input": copy.deepcopy(common.input_cfg_basic)}
+        min_mem_consump, max_mem_consump = JSON_checker.memory_consumption_estimation(
+            pipeline_cfg, input_cfg, pandora_machine
+        )
+        np.testing.assert_allclose((min_mem_consump, max_mem_consump), consumption_vt, rtol=1e-02)
 
 
 if __name__ == "__main__":