+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 ### Added
 
+- Memory consumption estimation. [#222]
 - Add extra requirements making mccnn plugin available.
 
 ## 1.1.0 (June 2021)